Extract common code in HTTP/2 tests
Follow the DRY principle.
