+++ dropwizard-http2/src/test/java/io/dropwizard/http2/Http2CIntegrationTest.java
@@ -1,16 +1,10 @@
 package io.dropwizard.http2;
 
-import com.google.common.base.Charsets;
 import com.google.common.net.HttpHeaders;
 import io.dropwizard.Configuration;
-import io.dropwizard.logging.BootstrapLogging;
 import io.dropwizard.testing.ResourceHelpers;
 import io.dropwizard.testing.junit.DropwizardAppRule;
 import org.eclipse.jetty.client.HttpClient;
-import org.eclipse.jetty.client.api.ContentResponse;
-import org.eclipse.jetty.client.api.Result;
-import org.eclipse.jetty.client.util.BufferingResponseListener;
-import org.eclipse.jetty.http.HttpVersion;
 import org.eclipse.jetty.http2.client.HTTP2Client;
 import org.eclipse.jetty.http2.client.HTTP2ClientConnectionFactory;
 import org.eclipse.jetty.http2.client.http.HttpClientTransportOverHTTP2;
@@ -23,16 +17,11 @@ import org.junit.Test;
 
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.TimeUnit;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class Http2CIntegrationTest {
+public class Http2CIntegrationTest  extends AbstractHttp2Test {
 
-    static {
-        BootstrapLogging.bootstrap();
-    }
 
     @Rule
     public DropwizardAppRule<Configuration> appRule = new DropwizardAppRule<>(
@@ -68,38 +57,14 @@ public class Http2CIntegrationTest {
 
     @Test
     public void testHttp2c() throws Exception {
-        final String hostname = "localhost";
-        final int port = appRule.getLocalPort();
-
-        final ContentResponse response = client.GET("http://" + hostname + ":" + port + "/api/test");
-        assertThat(response.getVersion()).isEqualTo(HttpVersion.HTTP_2);
-        assertThat(response.getStatus()).isEqualTo(200);
-        assertThat(response.getContentAsString()).isEqualTo(FakeApplication.HELLO_WORLD);
+        assertResponse(client.GET("http://localhost:" + appRule.getLocalPort() + "/api/test"));
     }
 
     @Test
     public void testHttp2cManyRequests() throws Exception {
-        final String hostname = "localhost";
-        final int port = appRule.getLocalPort();
-
         // For some reason the library requires to perform the first request synchronously with HTTP/2
         testHttp2c();
 
-        final int amount = 100;
-        final CountDownLatch latch = new CountDownLatch(amount);
-        for (int i = 0; i < amount; i++) {
-            client.newRequest("http://" + hostname + ":" + port + "/api/test")
-                    .send(new BufferingResponseListener() {
-                        @Override
-                        public void onComplete(Result result) {
-                            assertThat(result.getResponse().getVersion()).isEqualTo(HttpVersion.HTTP_2);
-                            assertThat(result.getResponse().getStatus()).isEqualTo(200);
-                            assertThat(getContentAsString(Charsets.UTF_8)).isEqualTo(FakeApplication.HELLO_WORLD);
-                            latch.countDown();
-                        }
-                    });
-        }
-
-        assertThat(latch.await(5, TimeUnit.SECONDS)).isTrue();
+        performManyAsyncRequests(client, "http://localhost:" + appRule.getLocalPort() + "/api/test");
     }
 }
